requirements.txt
tensorflow
streamlit
pillow
numpy
pandas
plotly
protobuf
